# swagger/swagger.yaml
basePath: /api/v1
definitions:
  erc-8004-benchmarking-be_internal_api_dto.Error:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  erc-8004-benchmarking-be_internal_api_dto.Meta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  erc-8004-benchmarking-be_internal_api_dto.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/erc-8004-benchmarking-be_internal_api_dto.Error'
      meta:
        $ref: '#/definitions/erc-8004-benchmarking-be_internal_api_dto.Meta'
      requestId:
        type: string
      success:
        type: boolean
    type: object
info:
  contact: {}
  description: REST API for leaderboard, agent profiles, OASF discovery, chains metadata,
    and admin observability.
  title: ERC-8004 Benchmarking API
  version: "1.0"
paths:
...
  /agents/{chainId}/{agentId}:
    get:
      parameters:
      - description: Chain ID
        in: path
        name: chainId
        required: true
        type: integer
      - description: Agent ID
        in: path
        name: agentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/erc-8004-benchmarking-be_internal_api_dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/erc-8004-benchmarking-be_internal_api_dto.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/erc-8004-benchmarking-be_internal_api_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/erc-8004-benchmarking-be_internal_api_dto.Response'
      summary: Agent profile
      tags:
      - agents
  /agents/{chainId}/{agentId}/activity-heatmap:
    get:
      parameters:
      - description: Chain ID
        in: path
        name: chainId
        required: true
        type: integer
      - description: Agent ID
        in: path
        name: agentId
        required: true
        type: string
      - description: Window in days
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/erc-8004-benchmarking-be_internal_api_dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/erc-8004-benchmarking-be_internal_api_dto.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/erc-8004-benchmarking-be_internal_api_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/erc-8004-benchmarking-be_internal_api_dto.Response'
      summary: Agent activity heatmap
      tags:
      - agents
  /agents/{chainId}/{agentId}/anti-spam:
    get:
      parameters:
      - description: Chain ID
        in: path
        name: chainId
        required: true
        type: integer
      - description: Agent ID
        in: path
        name: agentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/erc-8004-benchmarking-be_internal_api_dto.Response'
      summary: Agent anti-spam (v2 placeholder)
      tags:
      - agents
  /agents/{chainId}/{agentId}/feedbacks:
    get:
      parameters:
      - description: Chain ID
        in: path
        name: chainId
        required: true
        type: integer
      - description: Agent ID
        in: path
        name: agentId
        required: true
        type: string
      - description: Feedback category
        in: query
        name: category
        type: string
      - description: passed|failed|revoked
        in: query
        name: status
        type: string
      - description: RFC3339 start time
        in: query
        name: from
        type: string
      - description: RFC3339 end time
        in: query
        name: to
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      - description: newest|oldest
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/erc-8004-benchmarking-be_internal_api_dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/erc-8004-benchmarking-be_internal_api_dto.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/erc-8004-benchmarking-be_internal_api_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/erc-8004-benchmarking-be_internal_api_dto.Response'
      summary: Agent feedback list
      tags:
      - agents
  /agents/{chainId}/{agentId}/feedbacks/{feedbackId}:
    get:
      parameters:
      - description: Chain ID
        in: path
        name: chainId
        required: true
        type: integer
      - description: Agent ID
        in: path
        name: agentId
        required: true
        type: string
      - description: Feedback ID
        in: path
        name: feedbackId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/erc-8004-benchmarking-be_internal_api_dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/erc-8004-benchmarking-be_internal_api_dto.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/erc-8004-benchmarking-be_internal_api_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/erc-8004-benchmarking-be_internal_api_dto.Response'
      summary: Agent feedback detail
      tags:
      - agents
  /agents/{chainId}/{agentId}/identity-history:
    get:
      parameters:
      - description: Chain ID
        in: path
        name: chainId
        required: true
        type: integer
      - description: Agent ID
        in: path
        name: agentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/erc-8004-benchmarking-be_internal_api_dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/erc-8004-benchmarking-be_internal_api_dto.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/erc-8004-benchmarking-be_internal_api_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/erc-8004-benchmarking-be_internal_api_dto.Response'
      summary: Agent identity history
      tags:
      - agents
  /agents/{chainId}/{agentId}/penalties:
    get:
      parameters:
      - description: Chain ID
        in: path
        name: chainId
        required: true
        type: integer
      - description: Agent ID
        in: path
        name: agentId
        required: true
        type: string
      - description: fail|revoked|both
        in: query
        name: type
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/erc-8004-benchmarking-be_internal_api_dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/erc-8004-benchmarking-be_internal_api_dto.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/erc-8004-benchmarking-be_internal_api_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/erc-8004-benchmarking-be_internal_api_dto.Response'
      summary: Agent penalties
      tags:
      - agents
  /agents/{chainId}/{agentId}/proof/{txHash}:
    get:
      parameters:
      - description: Chain ID
        in: path
        name: chainId
        required: true
        type: integer
      - description: Agent ID
        in: path
        name: agentId
        required: true
        type: string
      - description: Transaction hash
        in: path
        name: txHash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/erc-8004-benchmarking-be_internal_api_dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/erc-8004-benchmarking-be_internal_api_dto.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/erc-8004-benchmarking-be_internal_api_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/erc-8004-benchmarking-be_internal_api_dto.Response'
      summary: Agent proof by transaction
      tags:
      - agents
  /agents/{chainId}/{agentId}/radar:
    get:
      parameters:
      - description: Chain ID
        in: path
        name: chainId
        required: true
        type: integer
      - description: Agent ID
        in: path
        name: agentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/erc-8004-benchmarking-be_internal_api_dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/erc-8004-benchmarking-be_internal_api_dto.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/erc-8004-benchmarking-be_internal_api_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/erc-8004-benchmarking-be_internal_api_dto.Response'
      summary: Agent radar
      tags:
      - agents
  /agents/{chainId}/{agentId}/related:
    get:
      parameters:
      - description: Chain ID
        in: path
        name: chainId
        required: true
        type: integer
      - description: Agent ID
        in: path
        name: agentId
        required: true
        type: string
      - description: skill|domain|both
        in: query
        name: by
        type: string
      - description: Max results
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/erc-8004-benchmarking-be_internal_api_dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/erc-8004-benchmarking-be_internal_api_dto.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/erc-8004-benchmarking-be_internal_api_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/erc-8004-benchmarking-be_internal_api_dto.Response'
      summary: Related agents
      tags:
      - agents
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"